feat(auth): implemented next-auth through custom form
